improve get_confirmations() rpc
Addresses #6 and #28

block_height.rs:
  * display BlockHeight as u64 to get rid of ugly leading zeros.

wallet_state.rs:
  * add block height to missing membership_proof warning message
  * remove get_balance_history()  (moved to wallet/mod.rs)

wallet/mod.rs:
  * log time taken in get_latest_balance_height().
  * rework get_latest_balance_height to find max(confirmed_in_block)
    ignoring any utxo not on current tip.
  * add get_balance_history() and rework to ignore any utxo not on
    current tip.  (issue #28)

lib.rs:
  * add 2 fn to time any fn call:
      time_fn_call() and time_fn_call_async()

rpc_server.rs:
  * call ::get_balance_history() on global state, not wallet_state.

+++ src/models/blockchain/block/block_height.rs
@@ -90,7 +90,7 @@ impl PartialOrd for BlockHeight {
 
 impl Display for BlockHeight {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
+        write!(f, "{}", u64::from(self.0))
     }
 }
